Move plugin struct definitions to model.go
